ci: Fix argo workflow name conflict
With the new release process, a push to main happens pretty much at the
same time as the tag generation. Notice that the same naming is used for
both workflows. This naming has a suffix that is set executing the
./scripts/version script, which will retrieve the current version with
the command:
  git describe --dirty --tags --long --always
If this command executed for the push CI is run after the tag is
created, it will generate a workflow name that conflicts with the one to
be created for the tag, not allowing the tag CI/CD to create its
workflow.

The fix adds an extra suffix to the workflow name that identifies the
workflow "mode", dev for pushes to main, release for tags.