Sync .github contents
Copy workflows from sm-alerts, where it's been tested that it's working.

We still need to pull the makefile changes. For now this will run in
parallel to the Drone set up.

This contains a number of fixes to get the build to work.

It also contains a change wrt to the original in order to get the 'test
docker build' step to work, as it needs to communicate with Docker, and
the socket belongs to different groups in different runners.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>