Fix TestHttpRunnerRun
With the current version it's possible to have a failure because the
randomly selected number is already used. That's not just bad luck
because the code doesn't try to recover from that.

Allow the OS to do its job, and simply ask what the outcome is. The
number is not needed for anything other than letting the client know
where to connect to.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>