Rework build_and_publish.yaml to handle docker images
Set up the global git config to disable ownership validation, this is
causing more grief than necesary and in this particular use case it
doesn't really provide anything. This is possibly a bug in
actions/checkout.

Tear shared-workflows/build-push-to-dockerhub appart and replace it with
a couple of steps (setup-buildx, build-push-action) in a way that the
image created in the buildx context is exported to a tar file, so that
we can reuse it in a different job to assemble a single multiarch image
together.

Add some intermediate steps that capture the result of building the
images and uploads them as build artifacts, so that we can download them
in a different job where the actual image is created.

Finally, put the images together and push them to GAR. This requires
reworking the dockerfiles in a way that they don't execute any platform
dependent steps, and everything can be done in the host.

Something broke with the chromium package, so I had to unpin it.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>