Fix: add packages to release
Use GitHub's command line tool (gh) to add packages to the existing
release. The assumption is that the tag was created because there's a
matching release.

This feels wrong, because we are creating a release and then later, in a
separate workflow, we are adding files to that release, instead of doing
it in a single location. From my point of view, the problem is that we
are conceptually creating a release which creates a tag instead of the
other way around. It's even a little weirder because GitHub does support
the concept of "draft release", but release-please doesn't.

Fixes: #956
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>