chore: Add support for chore commits in release-please (#1046)
* chore: Add support for chore commits in release-please

Include `chore` as a supported category for release-please. Chore
commits refer to changes that don't modify source files but are
necessary to maintain the project. Some examples include:
- Updating build scripts or configuration files
- Updating dependencies
- Updating CI/CD configurations

Consider using scopes in order to "narrow down" the context of the
change. For example:
- chore(ci): <change>
- chore(build): <change>

* Remove release commit type from release-please

The other supported categories should account for any change that may
require a release.