Fix: trigger argo release workflow from GHA
We have the drone pipeline running in parallel with the GHA one, with
the exception of the release step. In order to unblock this, disable
triggering the argo workflow in Drone and enable it in GHA. I'm 100%
sure this is going to fail without changes on the workflow side, but
making those changes first will very likely break the drone pipeline.
Start it here, with dev releases, so that we can figure out what else
needs changing but keep the prod path enabled in Drone, in case we need
to make a release before this is done.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>