Fix: pass version to argo workflow
We are passing the docker "tag", not just the version, and in this case
the tag includes information about the flavor of the image, which does
not match the version embedded in the binary.

As a footnote, yes, I know the tag is the full thing, not just the bits
after `:`, but that's how people refer to this. *shrug*

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>