fix: Update grafana-build-tools to v0.37.0
This pulls in Go 1.23.5, which contains a couple of security fixes, one
of them is relevant for our case.

The change is here: https://go-review.googlesource.com/c/go/+/643104

That change silently updates the test certificate used by httptest, and
that's being picked up by TestValidateMetrics because certificate's
"subjectalternative" field is changing. Regenerate the testdata.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>