fix: Use the recommended 'persist-credentials: false' setting (#1143)
The GitHub checkout action has this `persist-credentials` setting.

It's described like this:

    # Whether to configure the token or SSH key with the local git config
    # Default: true
    persist-credentials: ''

Sounds harmless, right? It almosts makes it sound as if it's going to
modify `~/.gitconfig`.

This default of `true` causes credentials to end up in the working
copy's `.git/config`, which interacts badly with other stuff happening
in GHA.

Set it to false just in case.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>