fix: Update Go version retrieval logic
`go list -m -f '{{.GoVersion}}'` works for standalone modules, but it
doesn't work in a Go workspace. In the later case, the command reports
the Go version of _all_ modules in the workspace.

`go mod edit -json` on the other hand only reports the versin in the
current module. Having to depend on jq like this is not ideal...

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>