fix: move handling of secrets store around
In the current version of the code, the secret store is populated when
the prober is created. The problem with that is that the prober is
created *once* in the lifetime of the check, but the secret store needs
to be updated more or less on a continuous basis (because it contains
the token used to access the secrets).

I'm sure there's a better way to handle this, but it took me a while to
figure out the issue.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>