fix: Synchronize makefiles with more recent versions (#1214)
We have multiple versions of these files floating around. Synchronize
the versions in this repo with the ones elsewhere, trying to avoid too
many changes at once.

This does remove the support for "bra", which was brought over from the
grafana repo, and as far as I know, nobody is actually using here.

Followup-to: 1982d52

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>