fix: Replace zerolog.Nop() with conditional logger
We have a few places where we use zerolog.Nop() in tests, which is fine
if we want to cut down the noise, but it's very difficult to use if
there are failures.

Change the direct calls to zerolog.Nop() to testhelper.Logger(t), which
will create a zerolog instance configured to filter out everything below
the error level. If the user passes the verbose flag to `go test`, when
it switches to using debug level.

Also add another helper, K6Path(*testing.T), which builds the path to
the k6 binary used in tests, and fails if that doesn't exist, to prevent
harder to debug errors down the test.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>