fix: Don't stop publisher if we hit Loki's stream limit (#1356)
Right now we are stopping the publisher if we hit Loki's stream limit,
as we consider that to be a fatal error. The actual observed behavior is
that Loki will accept *some* data and it will reject some. Even more
confusing, it seems like the stream was rejected (presumably the one
that is going over the limit), might be accepted at a later point in
time.

After this change the behavior is that whatever data was rejected is
dropped, and the rest of the publisher keeps working as if nothing had
happened. We will have to observe the actual behavior to decide if this
makes sense or not.

This is related to #1355.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>