fix: remove hardcoded blacklistIPs and blockHostnames for multihttp checks
`--blackist-ips` is already supplied as a CLI flag. Having it specified
here causes a conflict where this value will only take effect if
`--blacklist-ips` is empty, which is non-obvious.

`blockHostnames` follows the same pattern. On top of that, it's not
helping much: The default blocklist can be easily circumvented by not
fully qualifying a k8s hostname with `.cluster.local`, which will work
in most kubernetes environments due to the very common `ndots 5`
setting. Having this option here is more likely to create a sense of
false security than anything else.

Fixes https://github.com/grafana/synthetic-monitoring-agent/issues/735