feat: Implement usage report client
Implement usage report client that can send reports to https://stats.grafana.net/.
See https://github.com/grafana/usage-stats/ and https://github.com/grafana/usage-stats-handler/ for backend details about the stats endpoint.

Updates `checks.go` to submit a report whenever a probe is registered or updated.
The following metadata is sent over
- Version
- Arch
- OS
- Features
- Public bool
- TenantID

These are the fields that implemented as columns in usage stats(https://github.com/grafana/usage-stats/pull/55).
Any other fields that we add to the Report struct will be included as raw json and stored in a `report` field in BigQuery.

Added a flag `disable-usage-reports` which is initially set to true.
This provides a buffer so that we can announce that Grafana Labs will start to collect usage statistics for Agent's _and_ ensure documentation is clearly updated to reflect that usage reports will be gathered.