fix: runner: report script errors in the `RunResponse` object
The contract for the Runner is such that errors encountered while
running the script, as opposed to those encountered trying to start k6,
should be reported as part of the `RunResponse` object, using the
more-or-less human-readable `Error` and `ErrorCode` fields. If the `Run`
method returns an error, this is understood as a failure trying to run
k6 whatsoever, and no logs are reported to loki.

This commit fixes a bug where the result of `errorFromLogs`, which by
definition is always going to be an error with the script, was returned
as an error trying to start k6 instead of in the `Error` and `ErrorCode`
fields.