fix: error out early if grpc address doesn't have a port
For nearly all users, the port needs to be specified and needs to be
`443`. If skipped, GRPC will fall back to a different port that may not
respond to SYN TCP segments. As a result, the agent process will just
hang without logging anything.

This change causes the agent to error out if the port is omitted, which
allegedly should provide earlier feedback about this and other common
mistakes.