feat: Add CostAttributionLabels to checks proto
:warning: This is a highly experimental change that is mostly being used to validate a hypothethis on how cost attribution labels could work within the system.

Adds a `repeated Label costAttributionLabel` to the Check message. This is primarily meant to enable updating the API in such a way that I can
- add a CostAttributionLabel column to the checks table
- manually update local checks to add CostAttributionLabel stuff
- test out getting this data into the agent

Once I can validate that data propogates between the API and the agent, I will then work on updating the telemetry pipeline to start accounting for cost attribution labels.