Update telemeter to account for cost attribution labels
First take at attempting to account for cost attribution labels(CAL) in the telemeter package.
The naive approach is to extend the existing `CheckClassTelemetry` and add a map that keeps track of the execution times based upon `CostAttributionLabels` combinations in addition to the check class.

There are really four scenarios that need to be handled
1. no CAL set
2. label A set, label B not
3. label B set, label A not
4. label A and B set

The way I envision the keys is to map the label scenarios above with the keys below:

1. `unattributed`
2. `labelA=valueA`
3. `labelB=valueB`
4. `labelA=valueA&labelB=valueB`

This allows the telemeter package to track execution timings based upon every possible combination of CAL's.

When pushing values to the api, we iterate over all check classes and all the keys of CAL combinations and track their usage via the `TenantTelemetry` object.

While not implemented yet, the intention would be for the API to be able to the telemetry as it is sent over today. We would need to add an additional attribute on `CheckClassTelemetry` to inform the API what CAL is associated with the telemetry.

The main problem now is that we send multiple points of telemetry over to the API for each check. This current implementation would require the API to become aware of the changes, which I think is okay.