fix: Make tests less flaky
There are two separate issues here.

The first issue manifests itself as a data race in the logger. This is
known documented behavior of the test logger. The datarace is happening
happening because some goroutines are still running after the test is
done, and they try to log stuff. That code path *intentionally* ends up
in a datarace, so that `go test -race` can flag it. The message is not
the most obvious one, though.

This is addressed by throwing a waitgroup into the mix. The waitgroup is
not really going to do anything in production, but I don't think it's
going to hurt either. I was tempted to add a noop waitgroup
implementation, but decided against it because that requires introducing
an interface.

The second issue is that a couple of tests are trying to assert a
certain state that is the result of running goroutines near the end of
the test. Basically the issue is that you cannot expect goroutines to
run in a particular order or within some specific time frame. Without
some synchronization, it's simply not possible. And the code path that
is involved in this is intentionally avoiding synchronization.

This is addressed by simply retrying the assertion and sleeping a
little.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>