feat: Add IsUsingSecrets to synthetic_monitoring.Check
This method provides a single way to determine whether a check is using
secrets or not. For some types, this will be straighforward (either
because they do not support secrets, or because using secrets in that
case is very explicit). In other cases, this involves some guesswork,
and the extra scripts in this commit are mostly about validating that
guesswork.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>