fix: bbe config linter config
Removes the `//nolint:all` directive and fixes the linting errors.
There was a switch case that only had a single case and could be
rewritten as an if statement. I preferred this instead of adding another
`//nolint:...` directive.

Opted to add a `nolint:staticcheck` specific for the `s.Regexp`
condition as it wasn't obvious if `s.Regexp.String() ==
s.Regexp.Regexp.String()`.