fix: make log `duration_seconds` the same as the namesake metric
The agent defers into implementation-specific probers to report what is
the "latency" of whatever they're reporting. If they don't report one,
then the agent falls back to the wall time it took to run the prober.

Prior to this commit, that wall time was used unconditionally for
logging, but not for the metric, creating inconsistencies. This commit
changes that, so the reported value if used for both things if
available, and the wall time is also used for both things if not.

Additionally, the wall time is included as a separate attribute in the
logs, for completeness sake.

Closes https://github.com/grafana/synthetic-monitoring-agent/issues/1508