fix: Modernize the agent (#1503)
* fix: Modernize the agent

I get a solid amount of warnings from gopls or Goland when working
within the project, and they're often related to either newer features
being available or what I would consider nit's. An experimental tool
from Go exists called
[modernize](https://pkg.go.dev/golang.org/x/tools/gopls@v0.20.0/internal/analysis/modernize)
that allows you to automatically find and fix the code in a safe manner.
I was curious to see the diff after running it across the entire project
with the following command:

```
go run golang.org/x/tools/gopls/internal/analysis/modernize/cmd/modernize@latest -fix -test ./...
```

I've reviewed most of the changes and they all seem reasonably safe.
All unit tests appear to be passing as well.

* WIP

* fix: Whitespace linter breaking on tests

The modernizer tool removed the lines correctly but left blank spaces in
it's place, which golangci-linter is correctly complaining about.
Removes the blank spaces.

* fix: bbe config linter config

Removes the `//nolint:all` directive and fixes the linting errors.
There was a switch case that only had a single case and could be
rewritten as an if statement. I preferred this instead of adding another
`//nolint:...` directive.

Opted to add a `nolint:staticcheck` specific for the `s.Regexp`
condition as it wasn't obvious if `s.Regexp.String() ==
s.Regexp.Regexp.String()`.

* Temporarily add nolint:all to validate fixes

* Revert changes for bbe lint checks

Add a comment on why we do not want to lint the bbe files. These are
vendored in and we want to reduce the amount of drift between the
upstream implementation and synthetic-monitoring-agent.

* Rename maps0 to maps