chore: improve wording about how to disable default blocked networks
`--blocked-nets`, which controls the CIDR passed to k6's
[blocklist](https://grafana.com/docs/k6/latest/using-k6/k6-options/reference/#blacklist-ip),
defaults to `10.0.0.0/8`. This block can be lifted by setting
`--blocked-nets` to either the empty string, or a no-op value such as
`0.0.0.0/32`. This commit makes the help message more explicit on that,
as the previous could be misunderstood as "the default is no blocking".