feat: Add CostAttributionLabels to checks proto definition
First step in enabling cost attribution for agents within Synthetic
Monitoring. This first step updates the protobuf definition of a check
so that the changes can be brought into upstream repositories that will
propogate the changes across checks.

This does not update any of the code to handle cost attribution labels.
That will come in a future PR.

- refs https://github.com/grafana/synthetic-monitoring/issues/370