fix: Change default logging level to WarnLevel
Lowers the default severity level of logs which are outputted by the
application to Warn from Error.

This includes warning logs as a default item of interest for operators
and addresses an issue with adhoc checks on private probes wherein the
success message is not published to Loki due to its log level.