feat: Pull cost attribution labels for tenant
Following a similar path as limits to pull cost attribution labels for
tenants. Adds a new internal library `cals/tenant.go` that provides an
interface that wraps the `TenantProvider` client. Updates all of the
wiring from Updated => Scraper to pass along the new cal client.

Starts the process of using the library in `scraper.go`, which right now
finds the cost attribution labels for a specific tenant. Returns an
empty slice if nothing is found.

This will be dependant upon the API changes being released to get cost
attribution labels defined at the tenant level.