fix: Short circuit get tenant request
Use a default timeout for GetTenant call from agents to the API.
This takes especial relevance when calling GetTenant as part of a scrape
operation. In that context, a 5s TO for the fastest running protocol
checks, which have a min periodicity of 30s, seems a reasonable default
value considering the roundtrips between components in order to retrieve
the data.