feat: Adjust max tenant idle time
Specifies the time to wait for new check execution's data for the tenant
before removing the handler, which would cause any non delivered data to
be lost.

Adjust it to 65 minutes. In the worst case scenario, where the publisher
is unable to push for 1h, setting it to 65 minutes will allow the longer
running checks (current max=1h) to still buffer data for new executions.