feat: Increase max queued bytes
Specifies the max number of bytes to hold in the buffer per tenant in
case of unsuccessfull retries before discarding data.

In lack of better data for calculations.. running rough estimates we can
consider each check execution to report ~15 metrics with 10 labels
and one sample each. That will give us a rough estimate of ~5Kb per
check execution.

With the previous value of 128Kb, that would hold data for 25 check
executions. In the best scenario where a tenant runs a single check in
that agent, considering a check frequency of 1m, that would allow
for 25 minutes before discarding data. In most cases this number will be
much lower, as a tenant will be running more than one check in the same
agent, resulting in a very short time window before data is discarded.

Therefore, increase the buffer limit size to 1mb. Considering that that
will only have an effect when data is unable to be pushed to particular
Mimir/Loki cells, accept the memory increase in order to try to reduce
discarded data. Autoscalers will scale agents if necessary based on the
memory threshold.