feat: make `probe_duration_seconds` match `probe_script_duration_seconds`
For the vast majority of check types, `probe_duration_seconds` is
reported as the wall time that took to run the prober. This is accurate
for the vast majority of cases, but there's one in particular where it
is not: k6-backed checks that use external runners.

External runners can take a certain time to process a request, which
prior to this commit was counted as part of `probe_duration_seconds`
while in fact it was not. This commit adds code that parses the
`probe_script_duration_seconds` metric returned by k6 runners, and uses
that as the duration of the check for both logs and
`probe_duration_seconds`.