feat: Pull cost attribution labels for tenant (#1544)
Following a similar path as limits to pull cost attribution labels for
tenants. Adds a new internal library `cals/cals.go` that provides an
interface that wraps the `TenantProvider` client. Updates all of the
wiring from Updated => Scraper to pass along the new cal client.

Starts the process of using the library in `scraper.go`, which right now
finds the cost attribution labels for a specific tenant. Returns an
empty slice if nothing is found.

This will be dependent upon the API changes being released to get cost
attribution labels defined at the tenant level.

My goal is to have a follow up PR that updates the telemeter bits to track executions by cost attribution labels. 

- relies on https://github.com/grafana/synthetic-monitoring-api/pull/1750