WIP: Cal tracking refactoring
Still need to get all tests to pass.
I'm not convinced this is the route I'd like to take long term, but it
gets me incrementally closer.
One thing I need to consider is from the API's side, how do I ensure
that cal's not set are set to missing? This current implementation does
not handle that well. It also does not handle the case where there _are
no_ cal's.