Extend region_pusher AddExecution to track by CAL's
First attempt at extending the existing datastructures to account for
cost attribution labels.
Does not handle scenario's such as missing cost attribution, and the
unit tests still do not pass.

Adds an additional map to the region pushers `telemetry` attribute that
is responsible for tracking executions of cost attribution labels.
The `[]sm.CostAttributionLabels` need to be serialized in a way that
can be indexed on the map. `serializeCals` handles this by taking the
slice of cals and creates a comma separated value string.

This implementation does not track cal's by values just yet.