Track execution counts by cal names and values
Updates serialize/deserializeCals methods to account for the values as
well. The methods now take a slice of sm.CostAttributionLabels, which
are key/value pairs that allow the agent to keep track of the values
separately. This is done by converting the key/value pairs into `%s=%s`
format where the strings are the key and value respectively.
`__MISSING__` is used to track the case where no cals are set.

The region pusher will deserialize these back into costattributionlabels
so that the API can have access to the key/value pair without having to
perform any extra work.

Updates some tests to account for cost attribution labels. Work still
must be done to account for all four major scenarios.