feat: update CheckClassTelemetry protobuf (#1635)
* Update CheckClassTelemetry protobuf

Adds CostAttributionLabels field to the `CheckClassTelemetry` protobuf
and regenerates the protobufs. This is the first step of tracking
executions by cost attribution labels.

I'm making this minimal change so that we can release the change in the
agent and pull the changes into the API and validate no regressions.

* Update CheckClassTelemetry protobuf

After working through the change, I realized that we were not
communicating that values of teh cost attribution labels. We _could_
have serialized the values in such a
way(`[]string{"team=A","service=__missing__"}` and had the API
deserialize them, but this seems hacky.

This created a dedicated `CostAttributionLabel` message so we do not
tightly couple the existing `Label` to cost attribution labels.