feat: Resolve memcached servers (#1627)
In order to allow the user to pass a hostname on the command line (that
might resolve to multiple IP addresses), perform DNS resolution when
creating the Memcached client. The servers are resolved again on a
periodic basis to account for possible changes in IP addresses (as might
happen in a k8s cluster).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>