feat: track executions by cals (#1642)
Resolves https://github.com/grafana/synthetic-monitoring/issues/391

Extend the region pusher codebase to track executions by Cost Attribution Labels(CAL's) by adding an addition map to the telemetry attribution. The `telemetry` map now is now:

```go
// key int64: tenant id
// middle string: check class
// last string: serialized cost attribution label
telemetry   map[int64]map[sm.CheckClass]map[string]*sm.CheckClassTelemetry
```

CAL's are stored as `[]sm.CostAttributionLabel`, which can not be indexed in a map. Two methods were introduced:
1. `serializeCal`
2. `deserializeCal`

The methods allows the program to go from a slice to a string for the key and back. 

To preserve existing state, if no cost attribution labels are set, then the executions are tracked with the key `__MISSING__`. When executions are pushed to the API, this gets deserialized to `[]sm.CostAttributionLabel{}` which preserves the API's ability to emit the telemetry as it does today. A PR will be made to the API which will handle the scenarios where no cost attribution labels are set(existing state today), and when the slice is set.