docs: remove SECURITY.md
We have an org-wide SECURITY.md file that will be shown by GitHub
in the repository UI automatically. We are removing the grafana/grafana
copy of this file to avoid confusion and falling out of sync, so this
must also be updated to remove the reference to grafana/grafana's
security policy.

After this, this repository will simply inherit the org-wide policy, and
thus have that shown in the GitHub UI.